requirements.txt
aiohttp==3.8.4
fake_useragent==1.1.3
quart==0.18.4
googletrans==3.1.0a0
Requests==2.28.2
wsproto==1.0.0
quart_cors==0.6.0
jsonpickle==3.0.1
Flask-Cors==3.0.10
google-cloud-bigquery==2.26.0
google-auth==1.35.0

